refactor(builder): extract preamble helpers + split tests (dirge-4y4l stage 11a)
Begin decomposing the last god-file by ranking: src/agent/builder.rs
(1835 LOC). Convert the flat file into a builder/ directory module and
split out two concerns.

- preamble.rs (89 LOC) — system-prompt assembly text helpers:
  append_memory_to_preamble, assemble_base_preamble,
  model_steering_fragment, append_mode_reminder. Re-exported via
  `pub(crate) use preamble::*` (all are pub(crate)) so callers'
  paths are unchanged.
- reminder_tests.rs (689 LOC) — the #[cfg(test)] mod reminder_tests,
  moved verbatim. `use super::*` resolves through the re-exports.
- mod.rs (1068 LOC, was 1835) — still holds the two god-functions
  build_agent_inner (~516) and build_loop_tools (~425) plus
  build_session_search_tool / hookify / wrap_mcp_tools /
  DynamicToolSearch; those are stages 11b/11c. Dropped the prompt /
  model_family imports the moved helpers took with them.

Public API (build_agent_inner / build_loop_tools / wrap_mcp_tools /
DynamicToolSearch + the pub(crate) preamble helpers) preserved.
Behavior-preserving. -D warnings clean default + windows-default;
2374 tests green.

Refs dirge-553f
